// week08/Plat/mission2/src/types/auth.ts
import type { CommonResponse } from "./common";

//Signup
export type RequestSignupDto = {
    name: string;
    email: string;
    bio?: string;
    avatar?: string;
    password: string;
};

export type ResponseSignupDto = CommonResponse<{
    id: number;
    name: string;
    email: string;
    bio: string | null;
    avatar: string | null;
    createAt: Date;
    updateAt: Date;
}>;

//Login
export type RequestSigninDto = {
    email: string;
    password: string;
}

export type ResponseSigninDto = CommonResponse<{
    id: number;
    name: string;
    accessToken: string;
    refreshToken: string;
}>

// My Data Join
export type ResponseMyInfoDto = CommonResponse<{
    id: number;
    name: string;
    email: string;
    bio: string | null;
    avatar: string | null;
    createAt: Date;
    updateAt: Date;
}>

export type RequestUpdateMyInfoDto = {
    name: string;
    bio?: string | null;
    avatar?: string | null;
};

export type ResponseUpdateMyInfoDto = ResponseMyInfoDto;

export type ResponseLogoutDto = CommonResponse<null>;

export type ResponseDeleteMyAccountDto = CommonResponse<null>;